wip: lazy load placed marker in coordinates-input
