+++ src/components/ui/coordinates-input.tsx
@@ -8,10 +8,9 @@ import Map from "../map/map";
 import Button from "./button";
 import Dialog, { DialogSize, DialogTitleSize } from "./dialog";
 import TextField, { TTextFieldProps } from "./text-field";
-import { Dispatch, SetStateAction, useState } from "react";
-import { useMapEvent } from "react-leaflet";
-import MapMarker from "../map/map-marker";
+import { useState } from "react";
 import IconMarker from "./icons/marker";
+import dynamic from "next/dynamic";
 
 type TCoordinatesInputProps = TTextFieldProps & {
   dialogTitle: string;
@@ -93,7 +92,7 @@ function CoordinatesInput({
             zoom={mapZoom}
             scrollWheelZoom={true}
           >
-            <PlacedMarker
+            <LazyPlacedMarker
               position={markerPosition}
               setPosition={setMarkerPosition}
               markerType={markerType}
@@ -105,27 +104,9 @@ function CoordinatesInput({
   );
 }
 
-type TPlacedMarkerProps = {
-  position: [number, number] | undefined;
-  setPosition: Dispatch<SetStateAction<[number, number] | undefined>>;
-  markerType: "wall" | "parking";
-};
-
-function PlacedMarker({
-  position,
-  setPosition,
-  markerType,
-}: TPlacedMarkerProps) {
-  useMapEvent("click", (e) => {
-    setPosition([e.latlng.lat, e.latlng.lng]);
-  });
-
-  return (
-    position && (
-      <MapMarker type={markerType} position={position} interactive={false} />
-    )
-  );
-}
+export const LazyPlacedMarker = dynamic(() => import("./placed-marker"), {
+  ssr: false,
+});
 
 /**
  * Receives a string representation of coordinates
